+++ kuma/conftest.py
@@ -14,6 +14,15 @@ from kuma.wiki.constants import REDIRECT_CONTENT
 from kuma.wiki.models import Document, Revision
 
 
+@pytest.fixture
+def host_settings(settings):
+    settings.DOMAIN = 'mdn.dev'
+    settings.ALLOWED_HOSTS.append(settings.DOMAIN)
+    settings.ALLOWED_HOSTS.append(settings.BETA_HOST)
+    settings.ALLOWED_HOSTS.append(settings.WIKI_HOST)
+    return settings
+
+
 @pytest.fixture(autouse=True)
 def set_default_language():
     activate('en-US')
