+++ kuma/dashboards/urls.py
@@ -22,6 +22,10 @@ lang_urlpatterns = [
         views.topic_lookup,
         name='dashboards.topic_lookup'),
     url(r'^localization$',
+        # Here the "shared_cache_control" decorator is an optimization. It
+        # informs the CDN to cache the redirect for a week, so once this URL
+        # has been requested by a client, all other client requests will be
+        # redirected by the CDN instead of this Django service.
         shared_cache_control(s_maxage=WEEK)(RedirectView.as_view(
             url='/docs/MDN/Doc_status/Overview',
             permanent=True,
