+++ kuma/urls.py
@@ -47,6 +47,10 @@ urlpatterns += i18n_patterns(url('', include(landing_lang_urlpatterns)))
 urlpatterns += i18n_patterns(
     url(
         r'^events',
+        # Here the "shared_cache_control" decorator is an optimization. It
+        # informs the CDN to cache the redirect for a month, so once this URL
+        # has been requested by a client, all other client requests will be
+        # redirected by the CDN instead of this Django service.
         shared_cache_control(s_maxage=MONTH)(RedirectView.as_view(
             url='https://mozilla.org/contribute/events',
             permanent=False
