+++ kuma/wiki/jinja2/wiki/react_document.html
@@ -1,3 +1,4 @@
+// Set "wiki" and "untrusted" for the benefit of Jinja includes that need them.
 {% set wiki = False %}
 {% set untrusted = False %}
 {% set canonical = document_data['absoluteURL'] | absolutify %}
